fix(commands): Made changes related to pytest and its functions
